DSS does not load datafiles every time it is evaluated

+++ fragmentation/dss.cpp
@@ -19,7 +19,7 @@ extern "C"
         double &b, double &g);
 
     extern struct{
-        int fini;
+        double fini;
     } fragini_;
 }
